requirements.txt
requests>=2.31.0
pystray>=0.19.0
Pillow>=10.0.0
winotify>=1.1.0
